--- src/http/mod.rs
@@ -1,6 +1,0 @@
-pub mod controllers;
-/// This module contains functions references HTTP-related entities such as requests, responses, and routes.
-pub mod errors;
-pub mod health;
-mod middleware;
-pub mod openapi;
