+++ src/models/mod.rs
@@ -1,4 +1,2 @@
 /// This module contains all the models used in the application.
 pub mod api;
-pub mod identity_provider_registration;
-pub mod principal;
