+++ src/services/backends/kubernetes/identity_provider_repository/identity_provider_spec.rs
@@ -1,5 +1,5 @@
 use crate::models::api::external::identity_provider_settings::OidcExternalIdentityProviderSettings;
-use crate::models::identity_provider_registration::IdentityProviderRegistration;
+use crate::services::backends::kubernetes::identity_provider_repository::identity_provider_registration::IdentityProviderRegistration;
 use boxer_core::services::backends::kubernetes::kubernetes_repository::soft_delete_resource::SoftDeleteResource;
 use boxer_core::services::backends::kubernetes::kubernetes_repository::to_resource::ToResource;
 use boxer_core::services::backends::kubernetes::kubernetes_repository::try_from_resource::TryFromResource;
